+++ CHANGELOG.md
@@ -24,6 +24,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - IssuingToken resource
 - IssuingToken Log sub-resource
 - New attributes to issiuing resources
+- Ledger resource
+- LedgerTransaction resource
 
 ## [1.1.0] - 2026-05-18
 ### Added